creating 2 methods of data loading to database, performance comparison of each one, loading data to database with efficient method
